% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_deriv.R
\name{deriv2.logistic}
\alias{deriv2.logistic}
\title{Second Derivative}
\usage{
deriv2.logistic(X = NULL, Y = NULL, param = NULL)
}
\arguments{
\item{X}{is an n x p matrix of features}

\item{Y}{is an n x 1 vector of veatures}

\item{param}{is the parameter to be estimated by logistic()}
}
\description{
This function calculates the Hessian matrix for the logistic regression function.
}
