% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Logistic Regression}
\usage{
logistic(
  X = NULL,
  Y = NULL,
  start = NULL,
  tol = 1e-04,
  maxit = 50,
  verbose = 1
)
}
\arguments{
\item{X}{is an n x p data.matrix of features}

\item{Y}{is an n x 1 data.matrix of binary outcomes}

\item{start}{is the starting value}

\item{tol}{is the tolerance level (default is 1e-4)}

\item{maxit}{is the maximum iterations before ceasing maximization (default is 50)}

\item{verbose}{is the option to print iterated values (default is 1, 0 = no print)}
}
\description{
This function performs a logistic regression using the features in the data.
}
\examples{
\dontrun{
logistic(X,Y,tol=1e-5,maxit=10,verbose=0)
}
}
