% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_logistic.R
\name{predict.logistic}
\alias{predict.logistic}
\title{Logistic Regression Prediction}
\usage{
\method{predict}{logistic}(X = NULL, param = NULL)
}
\arguments{
\item{X}{is a n x p matrix with p features}

\item{param}{is the parameter estimated by the logistic() function}
}
\description{
This function uses logistic regression to predict features.
}
